refactor: move routes setup logic to composable
